requirements.txt
# Data fetching
requests==2.32.3
sodapy==2.2.0
python-dotenv==1.0.1

# Data processing
pandas==2.2.3

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.11

# API server
fastapi>=0.115.0
uvicorn>=0.32.0
apscheduler>=3.10.0
